# api/swagger.yaml
basePath: /api
definitions:
...
  portainer.AccessPolicy:
    properties:
      RoleId:
        description: Role identifier. Reference the role that will be associated to
          this access policy
        example: 1
        type: integer
    type: object
...
  portainer.GitlabRegistryData:
    properties:
      InstanceURL:
        type: string
      ProjectId:
        type: integer
      ProjectPath:
        type: string
    type: object
  portainer.EcrRegistryData:
    properties:
      AccessKeyID:
        type: string
      SecretAccessKey:
        type: string
      Region:
        type: string
    type: object
...
  portainer.Registry:
    properties:
      Authentication:
        description: Is authentication against this registry enabled
        example: true
        type: boolean
      AuthorizedTeams:
        items:
          type: integer
        type: array
      AuthorizedUsers:
        description: |-
          Deprecated fields
          Deprecated in DBVersion == 18
        items:
          description: User identifier who created this template
          example: 3
          type: integer
        type: array
      Ecr:
        $ref: '#/definitions/portainer.EcrRegistryData'
        description: AWS ECR specific details, required if Authentication is true and type = 6
      Gitlab:
        $ref: '#/definitions/portainer.GitlabRegistryData'
      Id:
        description: Registry Identifier
        example: 1
        type: integer
      ManagementConfiguration:
        $ref: '#/definitions/portainer.RegistryManagementConfiguration'
      Name:
        description: Registry Name
        example: my-registry
        type: string
      Password:
        description: Password used to authenticate against this registry
        example: registry_password
        type: string
      TeamAccessPolicies:
        $ref: '#/definitions/portainer.TeamAccessPolicies'
      Type:
        description: Registry Type (1 - Quay, 2 - Azure, 3 - Custom, 4 - Gitlab, 5 - ProGet, 6 - AWS ECR)
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        type: integer
      URL:
        description: URL or IP address of the Docker registry
        example: registry.mydomain.tld:2375
        type: string
      BaseURL:
        description: Base URL or IP address of the ProGet registry
        example: registry.mydomain.tld:2375
        type: string
      UserAccessPolicies:
        $ref: '#/definitions/portainer.UserAccessPolicies'
      Username:
        description: Username used to authenticate against this registry
        example: registry user
        type: string
    type: object
  portainer.RegistryManagementConfiguration:
    properties:
      Authentication:
        type: boolean
      Password:
        type: string
      TLSConfig:
        $ref: '#/definitions/portainer.TLSConfiguration'
      Type:
        type: integer
      Username:
        type: string
    type: object
...
  portainer.TLSConfiguration:
    properties:
      TLS:
        description: Use TLS
        example: true
        type: boolean
      TLSCACert:
        description: Path to the TLS CA certificate file
        example: /data/tls/ca.pem
        type: string
      TLSCert:
        description: Path to the TLS client certificate file
        example: /data/tls/cert.pem
        type: string
      TLSKey:
        description: Path to the TLS client key file
        example: /data/tls/key.pem
        type: string
      TLSSkipVerify:
        description: Skip the verification of the server TLS certificate
        example: false
        type: boolean
    type: object
...
  portainer.TeamAccessPolicies:
    additionalProperties:
      $ref: '#/definitions/portainer.AccessPolicy'
    type: object
...
  portainer.UserAccessPolicies:
    additionalProperties:
      $ref: '#/definitions/portainer.AccessPolicy'
    type: object
...
  registries.registryUpdatePayload:
    properties:
      authentication:
        description: Is authentication against this registry enabled
        example: false
        type: boolean
      ecr:
        $ref: '#/definitions/portainer.EcrRegistryData'
        description: AWS ECR specific details, required if authentication is true and type = 6
      name:
        description: Name that will be used to identify this registry
        example: my-registry
        type: string
      password:
        description: Password used to authenticate against this registry. required
          when Authentication is true
        example: registry_password
        type: string
      teamAccessPolicies:
        $ref: '#/definitions/portainer.TeamAccessPolicies'
      url:
        description: URL or IP address of the Docker registry
        example: registry.mydomain.tld:2375
        type: string
      baseUrl:
        description: Base URL or IP address of the ProGet registry
        example: registry.mydomain.tld:2375
        type: string
      userAccessPolicies:
        $ref: '#/definitions/portainer.UserAccessPolicies'
      username:
        description: Username used to authenticate against this registry. Required
          when Authentication is true
        example: registry_user
        type: string
    required:
    - authentication
    - name
    - url
    type: object
...
info:
  contact:
    email: info@portainer.io
  description: |
    Portainer API is an HTTP API served by Portainer. It is used by the Portainer UI and everything you can do with the UI can be done using the HTTP API.
    Examples are available at https://documentation.portainer.io/api/api-examples/
    You can find out more about Portainer at [http://portainer.io](http://portainer.io) and get some support on [Slack](http://portainer.io/slack/).

    # Authentication

    Most of the API endpoints require to be authenticated as well as some level of authorization to be used.
    Portainer API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:

    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOjEsImV4cCI6MTQ5OTM3NjE1NH0.NJ6vE8FY1WG6jsRQzfMqeatJ4vh2TWAeeYfDhP71YEE
    ```

    # Security

    Each API endpoint has an associated access policy, it is documented in the description of each endpoint.

    Different access policies are available:

    - Public access
    - Authenticated access
    - Restricted access
    - Administrator access

    ### Public access

    No authentication is required to access the endpoints with this access policy.

    ### Authenticated access

    Authentication is required to access the endpoints with this access policy.

...

...

...

...

...

...

...

...

...
paths:
...
  /registries/{id}:
    delete:
      description: |-
        Remove a registry
        **Access policy**: administrator
      operationId: RegistryDelete
      parameters:
      - description: Registry identifier
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Success
        "400":
          description: Invalid request
        "404":
          description: Registry not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Remove a registry
      tags:
      - registries
    get:
      description: |-
        Retrieve details about a registry.
        **Access policy**: administrator
      operationId: RegistryInspect
      parameters:
      - description: Registry identifier
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/portainer.Registry'
        "400":
          description: Invalid request
        "403":
          description: Permission denied to access registry
        "404":
          description: Registry not found
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Inspect a registry
      tags:
      - registries
    put:
      consumes:
      - application/json
      description: |-
        Update a registry
        **Access policy**: administrator
      operationId: RegistryUpdate
      parameters:
      - description: Registry identifier
        in: path
        name: id
        required: true
        type: integer
      - description: Registry details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/registries.registryUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/portainer.Registry'
        "400":
          description: Invalid request
        "404":
          description: Registry not found
        "409":
          description: Another registry with the same URL already exists
        "500":
          description: Server error
      security:
      - jwt: []
      summary: Update a registry
      tags:
      - registries
...
schemes:
...
securityDefinitions:
  jwt:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...